add static_casts to avoid unnecessary conversion from double to float

+++ Util/MercatorUtil.h
@@ -32,12 +32,12 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 inline float y2lat(const float a)
 {
-    return 180.f / M_PI * (2.f * std::atan(std::exp(a * M_PI / 180.f)) - M_PI / 2.f);
+    return 180.f / static_cast<float>(M_PI) * (2.f * std::atan(std::exp(a * static_cast<float>(M_PI) / 180.f)) - static_cast<float>(M_PI) / 2.f);
 }
 
 inline float lat2y(const float a)
 {
-    return 180.f / M_PI * std::log(std::tan(M_PI / 4.f + a * (M_PI / 180.f) / 2.f));
+    return 180.f / static_cast<float>(M_PI) * std::log(std::tan(static_cast<float>(M_PI) / 4.f + a * (static_cast<float>(M_PI) / 180.f) / 2.f));
 }
 
 #endif // MERCATOR_UTIL_H
